early access

+++ migrations/16-create-table-offer.sql
@@ -1,7 +1,7 @@
 CREATE TABLE offer(
     id serial NOT NULL,
     contract_address varchar,
-    price integer NOT NULL,
+    price bigint NOT NULL,
     date_created date NOT NULL,
     date_closed date,
     PRIMARY KEY (id)
